perf(seo): activate batched diagnostics projection
